better separation of pdf logic
