+++ src/Service/Report/IssueReport/IssueReportService.php
@@ -29,6 +29,11 @@ class IssueReportService implements IssueReportServiceInterface
      */
     private $translator;
 
+    /**
+     * IssueReportService constructor.
+     *
+     * @param TranslatorInterface $translator
+     */
     public function __construct(TranslatorInterface $translator)
     {
         $this->translator = $translator;
