+++ src/Service/Report/IssueReport/Pdf/Drawer.php
@@ -13,8 +13,8 @@ namespace App\Service\Report\IssueReport\Pdf;
 
 use App\Service\Report\Document\Transaction\Base\DrawableTransactionInterface;
 use App\Service\Report\IssueReport\Interfaces\DrawerInterface;
+use App\Service\Report\IssueReport\Pdf\Design\Interfaces\ColorServiceInterface;
 use App\Service\Report\Pdf\Cursor;
-use App\Service\Report\Pdf\Design\Interfaces\ColorServiceInterface;
 use App\Service\Report\Pdf\Interfaces\PdfDocumentInterface;
 
 class Drawer implements DrawerInterface
