+++ src/Service/ReportService.php
@@ -21,12 +21,25 @@ use App\Helper\IssueHelper;
 use App\Service\Interfaces\ImageServiceInterface;
 use App\Service\Interfaces\PathServiceInterface;
 use App\Service\Interfaces\ReportServiceInterface;
-use App\Service\Report\PdfDefinition;
-use App\Service\Report\Report;
+use App\Service\Report\Document\Interfaces\LayoutFactoryInterface;
+use App\Service\Report\IssueReport\Interfaces\IssueReportServiceInterface;
+use App\Service\Report\IssueReport\Interfaces\PrintFactoryInterface;
+use App\Service\Report\IssueReport\Model\AggregatedIssuesContent;
+use App\Service\Report\IssueReport\Model\IntroductionContent;
+use App\Service\Report\IssueReport\Model\IssueImage;
+use App\Service\Report\IssueReport\Model\MapContent;
+use App\Service\Report\IssueReport\Model\MetaData;
+use App\Service\Report\IssueReport\Pdf\Design\Interfaces\ColorServiceInterface;
+use App\Service\Report\IssueReport\Pdf\Design\Interfaces\LayoutServiceInterface;
+use App\Service\Report\IssueReport\Pdf\Design\Interfaces\TypographyServiceInterface;
+use App\Service\Report\IssueReport\PrintFactory;
+use App\Service\Report\Pdf\Interfaces\PageLayoutInterface;
+use App\Service\Report\Pdf\Interfaces\PdfDocumentInterface;
+use App\Service\Report\Pdf\Interfaces\PdfFactoryInterface;
+use App\Service\Report\Pdf\LayoutFactory;
 use App\Service\Report\ReportConfiguration;
 use App\Service\Report\ReportElements;
 use Symfony\Bridge\Doctrine\RegistryInterface;
-use Symfony\Component\Serializer\SerializerInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 class ReportService implements ReportServiceInterface
@@ -35,6 +48,21 @@ class ReportService implements ReportServiceInterface
      * @var PathServiceInterface
      */
     private $pathService;
+
+    /**
+     * @var LayoutServiceInterface
+     */
+    private $layoutService;
+
+    /**
+     * @var TypographyServiceInterface
+     */
+    private $typographyService;
+
+    /**
+     * @var ColorServiceInterface
+     */
+    private $colorService;
 
     /**
      * @var ImageServiceInterface
@@ -47,299 +75,44 @@ class ReportService implements ReportServiceInterface
     private $doctrine;
 
     /**
-     * @var SerializerInterface
-     */
-    private $serializer;
-
-    /**
      * @var TranslatorInterface
      */
     private $translator;
 
     /**
+     * @var PdfFactoryInterface
+     */
+    private $pdfFactory;
+
+    /**
+     * @var IssueReportServiceInterface
+     */
+    private $issueReportService;
+
+    /**
      * ReportService constructor.
      *
      * @param ImageServiceInterface $imageService
+     * @param LayoutServiceInterface $layoutService
      * @param RegistryInterface $registry
-     * @param SerializerInterface $serializer
      * @param TranslatorInterface $translator
      * @param PathServiceInterface $pathService
+     * @param PdfFactoryInterface $pdfFactory
+     * @param IssueReportServiceInterface $issueReportService
+     * @param TypographyServiceInterface $typographyService
+     * @param ColorServiceInterface $colorService
      */
-    public function __construct(ImageServiceInterface $imageService, RegistryInterface $registry, SerializerInterface $serializer, TranslatorInterface $translator, PathServiceInterface $pathService)
+    public function __construct(ImageServiceInterface $imageService, LayoutServiceInterface $layoutService, RegistryInterface $registry, TranslatorInterface $translator, PathServiceInterface $pathService, PdfFactoryInterface $pdfFactory, IssueReportServiceInterface $issueReportService, TypographyServiceInterface $typographyService, ColorServiceInterface $colorService)
     {
         $this->imageService = $imageService;
+        $this->layoutService = $layoutService;
         $this->doctrine = $registry;
-        $this->serializer = $serializer;
         $this->translator = $translator;
         $this->pathService = $pathService;
-    }
-
-    /**
-     * @param Report $report
-     * @param Map $map
-     * @param Issue[] $issues
-     */
-    private function addMap(Report $report, Map $map, array $issues)
-    {
-        $mapImage = $this->imageService->generateMapImageForReport($map, $issues, ImageServiceInterface::SIZE_REPORT_MAP);
-        if (file_exists($mapImage)) {
-            $report->addImage($map->getName(), $map->getContext(), $mapImage);
-        } else {
-            $report->addHeader($map->getName(), $map->getContext());
-        }
-    }
-
-    /**
-     * @param Report $report
-     * @param Issue[] $issues
-     */
-    private function addIssueImageGrid(Report $report, array $issues)
-    {
-        $columnCount = 4;
-
-        $imageGrid = [];
-        $currentRow = [];
-        foreach ($issues as $issue) {
-            $currentIssue = [];
-
-            $imagePath = $this->imageService->getSizeForIssue($issue, ImageServiceInterface::SIZE_REPORT_ISSUE);
-            if ($imagePath === null) {
-                continue;
-            }
-
-            $currentIssue['imagePath'] = $imagePath;
-            $currentIssue['identification'] = $issue->getNumber();
-            $currentRow[] = $currentIssue;
-
-            //add row to grid if applicable
-            if (\count($currentRow) === $columnCount) {
-                $imageGrid[] = $currentRow;
-                $currentRow = [];
-            }
-        }
-        if (\count($currentRow) > 0) {
-            $imageGrid[] = $currentRow;
-        }
-
-        $report->addImageGrid($imageGrid, $columnCount);
-    }
-
-    /**
-     * @param Report $report
-     * @param ConstructionSite $constructionSite
-     * @param Filter $filter
-     * @param ReportElements $reportElements
-     */
-    private function addIntroduction(Report $report, ConstructionSite $constructionSite, Filter $filter, ReportElements $reportElements)
-    {
-        $filterEntries = $this->getFilterEntries($filter);
-
-        //add list of elements which are part of this report
-        $elements = [];
-        if ($reportElements->getTableByCraftsman()) {
-            $elements[] = $this->translator->trans('table.by_craftsman', [], 'report');
-        }
-        if ($reportElements->getTableByMap()) {
-            $elements[] = $this->translator->trans('table.by_map', [], 'report');
-        }
-        if ($reportElements->getTableByTrade()) {
-            $elements[] = $this->translator->trans('table.by_trade', [], 'report');
-        }
-
-        $issueDetailsLabel = $this->translator->trans('issues.detailed', [], 'report');
-        if ($reportElements->getWithImages()) {
-            $issueDetailsLabel .= ' ' . $this->translator->trans('issues.with_images', [], 'report');
-        }
-        $elements[] = $issueDetailsLabel;
-
-        //print
-        $report->addIntroduction(
-            $this->imageService->getSizeForConstructionSite($constructionSite, ImageServiceInterface::SIZE_REPORT_ISSUE),
-            $constructionSite->getName(),
-            implode("\n", $constructionSite->getAddressLines()),
-            implode(', ', $elements),
-            $filterEntries,
-            $this->translator->trans('entity.name', [], 'entity_filter')
-        );
-    }
-
-    /**
-     * @param Report $report
-     * @param Issue[] $issues
-     * @param ReportConfiguration $reportConfiguration
-     */
-    private function addIssueTable(Report $report, array $issues, ReportConfiguration $reportConfiguration)
-    {
-        $tableHeader[] = '#';
-        $tableHeader[] = $this->translator->trans('description', [], 'entity_issue');
-        $tableHeader[] = $this->translator->trans('response_limit', [], 'entity_issue');
-
-        $tableContent = [];
-        foreach ($issues as $issue) {
-            $row = [];
-            $row[] = $issue->getNumber();
-            $row[] = $issue->getDescription();
-            $row[] = ($issue->getResponseLimit() !== null) ? $issue->getResponseLimit()->format(DateTimeFormatter::DATE_FORMAT) : '';
-            $tableContent[] = $row;
-        }
-
-        $issueCount = \count($issues);
-        if ($reportConfiguration->showRegistrationStatus()) {
-            $tableHeader[] = $this->translator->trans('table.in_state_since', ['%status%' => $this->translator->trans('status_values.registered', [], 'entity_issue')], 'report');
-            for ($i = 0; $i < $issueCount; ++$i) {
-                $issue = $issues[$i];
-                $tableContent[$i][] = $issue->getRegisteredAt() === null ? '' : $issue->getRegisteredAt()->format(DateTimeFormatter::DATE_FORMAT) . "\n" . $issue->getRegistrationBy()->getName();
-            }
-        }
-
-        if ($reportConfiguration->showRespondedStatus()) {
-            $tableHeader[] = $this->translator->trans('table.in_state_since', ['%status%' => $this->translator->trans('status_values.responded', [], 'entity_issue')], 'report');
-            for ($i = 0; $i < $issueCount; ++$i) {
-                $issue = $issues[$i];
-                $tableContent[$i][] = $issue->getRespondedAt() === null ? '' : $issue->getRespondedAt()->format(DateTimeFormatter::DATE_FORMAT) . "\n" . $issue->getResponseBy()->getName();
-            }
-        }
-
-        if ($reportConfiguration->showReviewedStatus()) {
-            $tableHeader[] = $this->translator->trans('table.in_state_since', ['%status%' => $this->translator->trans('status_values.reviewed', [], 'entity_issue')], 'report');
-            for ($i = 0; $i < $issueCount; ++$i) {
-                $issue = $issues[$i];
-                $tableContent[$i][] = $issue->getReviewedAt() === null ? '' : $issue->getReviewedAt()->format(DateTimeFormatter::DATE_FORMAT) . "\n" . $issue->getReviewBy()->getName();
-            }
-        }
-
-        $report->addTable($tableHeader, $tableContent);
-    }
-
-    /**
-     * @param array $orderedMaps
-     * @param Issue[][] $issuesPerMap
-     * @param array $tableContent
-     * @param array $tableHeader
-     * @param ReportConfiguration $configuration
-     */
-    private function addAggregatedIssuesInfo(array $orderedMaps, array $issuesPerMap, array &$tableContent, array &$tableHeader, ReportConfiguration $configuration)
-    {
-        //count issue status per map
-        $countsPerElement = [];
-        foreach ($orderedMaps as $index => $element) {
-            $countPerMap = [0, 0, 0];
-            foreach ($issuesPerMap[$index] as $issue) {
-                if ($issue->getStatusCode() >= Issue::REVIEW_STATUS) {
-                    ++$countPerMap[2];
-                } elseif ($issue->getStatusCode() >= Issue::RESPONSE_STATUS) {
-                    ++$countPerMap[1];
-                } else {
-                    ++$countPerMap[0];
-                }
-            }
-            $countsPerElement[$index] = $countPerMap;
-        }
-
-        //add registration count if filter did not exclude
-        if ($configuration->showRegistrationStatus()) {
-            $tableHeader[] = $this->translator->trans('status_values.registered', [], 'entity_issue');
-            foreach ($countsPerElement as $elementId => $count) {
-                $tableContent[$elementId][] = $count[0];
-            }
-        }
-
-        //add response count if filter did not exclude
-        if ($configuration->showRespondedStatus()) {
-            $tableHeader[] = $this->translator->trans('status_values.responded', [], 'entity_issue');
-            foreach ($countsPerElement as $elementId => $count) {
-                $tableContent[$elementId][] = $count[1];
-            }
-        }
-
-        //add review count if filter did not exclude
-        if ($configuration->showReviewedStatus()) {
-            $tableHeader[] = $this->translator->trans('status_values.reviewed', [], 'entity_issue');
-            foreach ($countsPerElement as $elementId => $count) {
-                $tableContent[$elementId][] = $count[2];
-            }
-        }
-    }
-
-    /**
-     * @param Report $report
-     * @param Issue[] $issues
-     * @param ReportConfiguration $reportConfiguration
-     */
-    private function addTableByMap(Report $report, array $issues, ReportConfiguration $reportConfiguration)
-    {
-        /* @var Map[] $orderedMaps */
-        /* @var Issue[][] $issuesPerMap */
-        IssueHelper::issuesToOrderedMaps($issues, $orderedMaps, $issuesPerMap);
-
-        //prepare header & content with specific content
-        $tableHeader = [$this->translator->trans('context', [], 'entity_map'), $this->translator->trans('entity.name', [], 'entity_map')];
-
-        //add map name & map context to table
-        $tableContent = [];
-        foreach ($orderedMaps as $mapId => $map) {
-            $tableContent[$mapId] = [$map->getContext(), $map->getName()];
-        }
-
-        //add accumulated info
-        $this->addAggregatedIssuesInfo($orderedMaps, $issuesPerMap, $tableContent, $tableHeader, $reportConfiguration);
-
-        //write to pdf
-        $report->addTable($tableHeader, $tableContent, $this->translator->trans('table.by_map', [], 'report'));
-    }
-
-    /**
-     * @param Report $report
-     * @param Issue[] $issues
-     * @param ReportConfiguration $reportConfiguration
-     */
-    private function addTableByCraftsman(Report $report, array $issues, ReportConfiguration $reportConfiguration)
-    {
-        /* @var Craftsman[] $orderedCraftsman */
-        /* @var Issue[][] $issuesPerCraftsman */
-        IssueHelper::issuesToOrderedCraftsman($issues, $orderedCraftsman, $issuesPerCraftsman);
-
-        //prepare header & content with specific content
-        $tableHeader = [$this->translator->trans('entity.name', [], 'entity_craftsman')];
-
-        //add map name & map context to table
-        $tableContent = [];
-        foreach ($orderedCraftsman as $craftsmanId => $craftsman) {
-            $tableContent[$craftsmanId] = [$craftsman->getName()];
-        }
-
-        //add accumulated info
-        $this->addAggregatedIssuesInfo($orderedCraftsman, $issuesPerCraftsman, $tableContent, $tableHeader, $reportConfiguration);
-
-        //write to pdf
-        $report->addTable($tableHeader, $tableContent, $this->translator->trans('table.by_craftsman', [], 'report'));
-    }
-
-    /**
-     * @param Report $report
-     * @param Issue[] $issues
-     * @param ReportConfiguration $reportConfiguration
-     */
-    private function addTableByTrade(Report $report, array $issues, ReportConfiguration $reportConfiguration)
-    {
-        /* @var string[] $orderedTrade */
-        /* @var Issue[][] $issuesPerTrade */
-        IssueHelper::issuesToOrderedTrade($issues, $orderedTrade, $issuesPerTrade);
-
-        //prepare header & content with specific content
-        $tableHeader = [$this->translator->trans('trade', [], 'entity_craftsman')];
-
-        //add map name & map context to table
-        $tableContent = [];
-        foreach ($orderedTrade as $trade) {
-            $tableContent[$trade] = [$trade];
-        }
-
-        //add accumulated info
-        $this->addAggregatedIssuesInfo($orderedTrade, $issuesPerTrade, $tableContent, $tableHeader, $reportConfiguration);
-
-        //write to pdf
-        $report->addTable($tableHeader, $tableContent, $this->translator->trans('table.by_trade', [], 'report'));
+        $this->pdfFactory = $pdfFactory;
+        $this->issueReportService = $issueReportService;
+        $this->typographyService = $typographyService;
+        $this->colorService = $colorService;
     }
 
     /**
@@ -354,25 +127,52 @@ class ReportService implements ReportServiceInterface
      */
     public function generatePdfReport(ConstructionSite $constructionSite, Filter $filter, string $author, ReportElements $reportElements)
     {
+        // initialize pdf report
+        $printFactory = new PrintFactory($this->typographyService, $this->colorService, $this->layoutService);
+
+        $pageLayoutContent = $this->getMetaData($constructionSite, $author);
+        $layout = $printFactory->getLayout($pageLayoutContent);
+        $document = $this->createPdfDocument($layout);
+
+        $layoutFactory = new LayoutFactory($document, $this->layoutService);
+        $this->addReportElements($layoutFactory, $printFactory, $constructionSite, $filter, $reportElements);
+
+        // persist
+        $filePath = $this->getFilePath($constructionSite);
+        $document->save($filePath);
+
+        return $filePath;
+    }
+
+    /**
+     * @param LayoutFactoryInterface $layoutFactory
+     * @param PrintFactoryInterface $buildingBlocks
+     * @param ConstructionSite $constructionSite
+     * @param Filter $filter
+     * @param ReportElements $reportElements
+     */
+    private function addReportElements(LayoutFactoryInterface $layoutFactory, PrintFactoryInterface $buildingBlocks, ConstructionSite $constructionSite, Filter $filter, ReportElements $reportElements)
+    {
         $issues = $this->doctrine->getRepository(Issue::class)->filter($filter);
         $reportConfiguration = new ReportConfiguration($filter);
 
-        // initialize report
-        $footnote = $this->translator->trans('generated', ['%date%' => (new \DateTime())->format(DateTimeFormatter::DATE_TIME_FORMAT), '%name%' => $author], 'report');
-        $pdfDefinition = new PdfDefinition($constructionSite->getName(), $footnote, __DIR__ . '/../../assets/report/logo.png');
-        $report = new Report($pdfDefinition);
-
-        $this->addIntroduction($report, $constructionSite, $filter, $reportElements);
+        // add introduction
+        $introductionContent = $this->getIntroductionContent($constructionSite, $filter, $reportElements);
+        $this->issueReportService->addIntroduction($layoutFactory, $buildingBlocks, $introductionContent);
 
         //add tables
+        $aggregatedIssues = [];
         if ($reportElements->getTableByCraftsman()) {
-            $this->addTableByCraftsman($report, $issues, $reportConfiguration);
+            $aggregatedIssues[] = $this->aggregateIssuesByCraftsman($issues, $reportConfiguration);
         }
         if ($reportElements->getTableByMap()) {
-            $this->addTableByMap($report, $issues, $reportConfiguration);
+            $aggregatedIssues[] = $this->aggregatedIssuesByMap($issues, $reportConfiguration);
         }
         if ($reportElements->getTableByTrade()) {
-            $this->addTableByTrade($report, $issues, $reportConfiguration);
+            $aggregatedIssues[] = $this->aggregatedIssuesByTrade($issues, $reportConfiguration);
+        }
+        foreach ($aggregatedIssues as $aggregatedIssue) {
+            $this->issueReportService->addAggregatedIssueTable($layoutFactory, $buildingBlocks, $aggregatedIssue);
         }
 
         /* @var Map[] $orderedMaps */
@@ -381,18 +181,340 @@ class ReportService implements ReportServiceInterface
         foreach ($orderedMaps as $map) {
             $issues = $issuesPerMap[$map->getId()];
 
-            $this->addMap($report, $map, $issues);
-            $this->addIssueTable($report, $issues, $reportConfiguration);
+            $mapContent = $this->getMapContent($map, $issues, $reportElements, $reportConfiguration);
+            $this->issueReportService->addMap($layoutFactory, $buildingBlocks, $mapContent);
+        }
+    }
 
-            if ($reportElements->getWithImages()) {
-                $this->addIssueImageGrid($report, $issues);
+    /**
+     * @param PageLayoutInterface $pageLayout
+     *
+     * @return PdfDocumentInterface
+     */
+    private function createPdfDocument(PageLayoutInterface $pageLayout)
+    {
+        $fontPath = $this->pathService->getAssetsRoot() . \DIRECTORY_SEPARATOR . 'report' . \DIRECTORY_SEPARATOR . 'fonts';
+        $this->pdfFactory->configure(['tcpdf' => ['font_path' => $fontPath]]);
+
+        return $this->pdfFactory->create($pageLayout);
+    }
+
+    /**
+     * @param ConstructionSite $constructionSite
+     * @param string $author
+     *
+     * @throws \Exception
+     *
+     * @return MetaData
+     */
+    private function getMetaData(ConstructionSite $constructionSite, string $author)
+    {
+        $pageLayoutContent = new MetaData();
+
+        $pageLayoutContent->setTitle($constructionSite->getName());
+        $pageLayoutContent->setAuthor($author);
+
+        $formattedDateTime = (new \DateTime())->format(DateTimeFormatter::DATE_TIME_FORMAT);
+        $pageLayoutContent->setGenerationInfoText($this->translator->trans('generated', ['%date%' => $formattedDateTime, '%name%' => $author], 'report'));
+
+        $logoPath = $this->pathService->getAssetsRoot() . \DIRECTORY_SEPARATOR . 'report' . \DIRECTORY_SEPARATOR . 'logo.png';
+        $pageLayoutContent->setLogoPath($logoPath);
+
+        return $pageLayoutContent;
+    }
+
+    /**
+     * @param ConstructionSite $constructionSite
+     * @param Filter $filter
+     * @param ReportElements $reportElements
+     *
+     * @return IntroductionContent
+     */
+    private function getIntroductionContent(ConstructionSite $constructionSite, Filter $filter, ReportElements $reportElements)
+    {
+        $introductionContent = new IntroductionContent();
+        $introductionContent->setConstructionSiteName($constructionSite->getName());
+        $introductionContent->setConstructionSiteImage($this->imageService->getSizeForConstructionSite($constructionSite, ImageServiceInterface::SIZE_REPORT_ISSUE));
+        $introductionContent->setConstructionSiteAddressLines($constructionSite->getAddressLines());
+        $introductionContent->setFilterEntries($this->getFilterEntries($filter));
+        $introductionContent->setReportElements($this->getReportElements($reportElements));
+
+        return $introductionContent;
+    }
+
+    /**
+     * @param Issue[] $issues
+     * @param ReportConfiguration $reportConfiguration
+     *
+     * @return AggregatedIssuesContent
+     */
+    private function aggregateIssuesByCraftsman(array $issues, ReportConfiguration $reportConfiguration)
+    {
+        $aggregatedIssues = new AggregatedIssuesContent();
+        $aggregatedIssues->setTableDescription($this->translator->trans('table.by_craftsman', [], 'report'));
+
+        /* @var Craftsman[] $orderedCraftsman */
+        /* @var Issue[][] $issuesPerCraftsman */
+        IssueHelper::issuesToOrderedCraftsman($issues, $orderedCraftsman, $issuesPerCraftsman);
+
+        //prepare header & content with specific content
+        $aggregatedIssues->setIdentifierHeader([$this->translator->trans('entity.name', [], 'entity_craftsman')]);
+
+        //add name table
+        foreach ($orderedCraftsman as $craftsmanId => $craftsman) {
+            $aggregatedIssues->addIdentifierContent([$craftsman->getName()]);
+        }
+
+        $aggregatedIssues->setIssuesHeader($this->getAggregatedIssuesTableHeader($reportConfiguration));
+        $aggregatedIssues->setIssuesContent($this->getAggregatedIssuesTableContent($orderedCraftsman, $issuesPerCraftsman, $reportConfiguration));
+
+        return $aggregatedIssues;
+    }
+
+    /**
+     * @param Issue[] $issues
+     * @param ReportConfiguration $reportConfiguration
+     *
+     * @return AggregatedIssuesContent
+     */
+    private function aggregatedIssuesByMap(array $issues, ReportConfiguration $reportConfiguration)
+    {
+        $aggregatedIssues = new AggregatedIssuesContent();
+        $aggregatedIssues->setTableDescription($this->translator->trans('table.by_map', [], 'report'));
+
+        /* @var Map[] $orderedMaps */
+        /* @var Issue[][] $issuesPerMap */
+        IssueHelper::issuesToOrderedMaps($issues, $orderedMaps, $issuesPerMap);
+
+        //prepare header & content with specific content
+        $aggregatedIssues->setIdentifierHeader([$this->translator->trans('context', [], 'entity_map'), $this->translator->trans('entity.name', [], 'entity_map')]);
+
+        //add name table
+        foreach ($orderedMaps as $mapId => $map) {
+            $aggregatedIssues->addIdentifierContent([$map->getContext(), $map->getName()]);
+        }
+
+        $aggregatedIssues->setIssuesHeader($this->getAggregatedIssuesTableHeader($reportConfiguration));
+        $aggregatedIssues->setIssuesContent($this->getAggregatedIssuesTableContent($orderedMaps, $issuesPerMap, $reportConfiguration));
+
+        return $aggregatedIssues;
+    }
+
+    /**
+     * @param Issue[] $issues
+     * @param ReportConfiguration $reportConfiguration
+     *
+     * @return AggregatedIssuesContent
+     */
+    private function aggregatedIssuesByTrade(array $issues, ReportConfiguration $reportConfiguration)
+    {
+        $aggregatedIssues = new AggregatedIssuesContent();
+        $aggregatedIssues->setTableDescription($this->translator->trans('table.by_trade', [], 'report'));
+
+        /* @var string[] $orderedTrade */
+        /* @var Issue[][] $issuesPerTrade */
+        IssueHelper::issuesToOrderedTrade($issues, $orderedTrade, $issuesPerTrade);
+
+        //prepare header & content with specific content
+        $aggregatedIssues->setIdentifierHeader([$this->translator->trans('trade', [], 'entity_craftsman')]);
+
+        //add name table
+        foreach ($orderedTrade as $trade) {
+            $aggregatedIssues->addIdentifierContent([$trade]);
+        }
+
+        $aggregatedIssues->setIssuesHeader($this->getAggregatedIssuesTableHeader($reportConfiguration));
+        $aggregatedIssues->setIssuesContent($this->getAggregatedIssuesTableContent($orderedTrade, $issuesPerTrade, $reportConfiguration));
+
+        return $aggregatedIssues;
+    }
+
+    /**
+     * @param ReportConfiguration $configuration
+     *
+     * @return string[]
+     */
+    private function getAggregatedIssuesTableHeader(ReportConfiguration $configuration)
+    {
+        $tableHeader = [];
+
+        //add registration count if filter did not exclude
+        if ($configuration->showRegistrationStatus()) {
+            $tableHeader[] = $this->translator->trans('status_values.registered', [], 'entity_issue');
+        }
+
+        //add response count if filter did not exclude
+        if ($configuration->showRespondedStatus()) {
+            $tableHeader[] = $this->translator->trans('status_values.responded', [], 'entity_issue');
+        }
+
+        //add review count if filter did not exclude
+        if ($configuration->showReviewedStatus()) {
+            $tableHeader[] = $this->translator->trans('status_values.reviewed', [], 'entity_issue');
+        }
+
+        return $tableHeader;
+    }
+
+    /**
+     * @param array $elements
+     * @param Issue[][] $issuesPerElement
+     * @param ReportConfiguration $configuration
+     *
+     * @return string[][]
+     */
+    private function getAggregatedIssuesTableContent(array $elements, array $issuesPerElement, ReportConfiguration $configuration)
+    {
+        $tableContent = [];
+
+        //count issue status per map
+        $countPerElements = [];
+        foreach ($elements as $index => $element) {
+            $countPerElement = [0, 0, 0];
+            foreach ($issuesPerElement[$index] as $issue) {
+                if ($issue->getStatusCode() >= Issue::REVIEW_STATUS) {
+                    ++$countPerElement[2];
+                } elseif ($issue->getStatusCode() >= Issue::RESPONSE_STATUS) {
+                    ++$countPerElement[1];
+                } else {
+                    ++$countPerElement[0];
+                }
+            }
+            $countPerElements[$index] = $countPerElement;
+        }
+
+        //add registration count if filter did not exclude
+        if ($configuration->showRegistrationStatus()) {
+            foreach ($countPerElements as $elementId => $count) {
+                $tableContent[$elementId][] = $count[0];
             }
         }
 
-        $filePath = $this->getFilePath($constructionSite);
-        $report->save($filePath);
+        //add response count if filter did not exclude
+        if ($configuration->showRespondedStatus()) {
+            foreach ($countPerElements as $elementId => $count) {
+                $tableContent[$elementId][] = $count[1];
+            }
+        }
 
-        return $filePath;
+        //add review count if filter did not exclude
+        if ($configuration->showReviewedStatus()) {
+            foreach ($countPerElements as $elementId => $count) {
+                $tableContent[$elementId][] = $count[2];
+            }
+        }
+
+        return array_values($tableContent);
+    }
+
+    /**
+     * @param Map $map
+     * @param Issue[] $issues
+     * @param ReportElements $elements
+     * @param ReportConfiguration $reportConfiguration
+     *
+     * @return MapContent
+     */
+    private function getMapContent(Map $map, array $issues, ReportElements $elements, ReportConfiguration $reportConfiguration)
+    {
+        $mapContent = new MapContent();
+
+        $mapContent->setMapName($map->getName());
+        $mapContent->setMapContext($map->getContext());
+        $mapContent->setMapImage($this->imageService->generateMapImageForReport($map, $issues, ImageServiceInterface::SIZE_REPORT_MAP));
+
+        $mapContent->setIssuesTableHeader($this->getIssuesTableHeader($reportConfiguration));
+        $mapContent->setIssuesTableContent($this->getIssuesTableContent($issues, $reportConfiguration));
+
+        if ($elements->getWithImages()) {
+            foreach ($this->getIssueImages($issues) as $issueImage) {
+                $mapContent->addIssueImage($issueImage);
+            }
+        }
+
+        return $mapContent;
+    }
+
+    /**
+     * @param Issue[] $issues
+     *
+     * @return IssueImage[]
+     */
+    private function getIssueImages(array $issues)
+    {
+        $issueImages = [];
+        foreach ($issues as $issue) {
+            $imagePath = $this->imageService->getSizeForIssue($issue, ImageServiceInterface::SIZE_REPORT_ISSUE);
+            if ($imagePath === null) {
+                continue;
+            }
+
+            $issueImage = new IssueImage();
+            $issueImage->setImagePath($imagePath);
+            $issueImage->setNumber($issue->getNumber());
+            $issueImages[] = $issueImage;
+        }
+
+        return $issueImages;
+    }
+
+    /**
+     * @param Issue[] $issues
+     * @param ReportConfiguration $reportConfiguration
+     *
+     * @return string[][]
+     */
+    private function getIssuesTableContent(array $issues, ReportConfiguration $reportConfiguration)
+    {
+        $tableContent = [];
+        foreach ($issues as $issue) {
+            $row = [];
+            $row[] = $issue->getNumber();
+            $row[] = $issue->getDescription();
+            $row[] = ($issue->getResponseLimit() !== null) ? $issue->getResponseLimit()->format(DateTimeFormatter::DATE_FORMAT) : '';
+
+            if ($reportConfiguration->showRegistrationStatus()) {
+                $row[] = $issue->getRegisteredAt() === null ? '' : $issue->getRegisteredAt()->format(DateTimeFormatter::DATE_FORMAT) . "\n" . $issue->getRegistrationBy()->getName();
+            }
+
+            if ($reportConfiguration->showRespondedStatus()) {
+                $row[] = $issue->getRespondedAt() === null ? '' : $issue->getRespondedAt()->format(DateTimeFormatter::DATE_FORMAT) . "\n" . $issue->getResponseBy()->getName();
+            }
+
+            if ($reportConfiguration->showReviewedStatus()) {
+                $row[] = $issue->getReviewedAt() === null ? '' : $issue->getReviewedAt()->format(DateTimeFormatter::DATE_FORMAT) . "\n" . $issue->getReviewBy()->getName();
+            }
+
+            $tableContent[] = $row;
+        }
+
+        return $tableContent;
+    }
+
+    /**
+     * @param ReportConfiguration $reportConfiguration
+     *
+     * @return array
+     */
+    private function getIssuesTableHeader(ReportConfiguration $reportConfiguration)
+    {
+        $tableHeader[] = '#';
+        $tableHeader[] = $this->translator->trans('description', [], 'entity_issue');
+        $tableHeader[] = $this->translator->trans('response_limit', [], 'entity_issue');
+
+        if ($reportConfiguration->showRegistrationStatus()) {
+            $tableHeader[] = $this->translator->trans('table.in_state_since', ['%status%' => $this->translator->trans('status_values.registered', [], 'entity_issue')], 'report');
+        }
+
+        if ($reportConfiguration->showRespondedStatus()) {
+            $tableHeader[] = $this->translator->trans('table.in_state_since', ['%status%' => $this->translator->trans('status_values.responded', [], 'entity_issue')], 'report');
+        }
+
+        if ($reportConfiguration->showReviewedStatus()) {
+            $tableHeader[] = $this->translator->trans('table.in_state_since', ['%status%' => $this->translator->trans('status_values.reviewed', [], 'entity_issue')], 'report');
+        }
+
+        return $tableHeader;
     }
 
     /**
@@ -558,5 +680,31 @@ class ReportService implements ReportServiceInterface
         }
 
         return '';
+    }
+
+    /**
+     * @param ReportElements $reportElements
+     *
+     * @return string[]
+     */
+    private function getReportElements(ReportElements $reportElements): array
+    {
+        if ($reportElements->getTableByCraftsman()) {
+            $elements[] = $this->translator->trans('table.by_craftsman', [], 'report');
+        }
+        if ($reportElements->getTableByMap()) {
+            $elements[] = $this->translator->trans('table.by_map', [], 'report');
+        }
+        if ($reportElements->getTableByTrade()) {
+            $elements[] = $this->translator->trans('table.by_trade', [], 'report');
+        }
+
+        $issueDetailsLabel = $this->translator->trans('issues.detailed', [], 'report');
+        if ($reportElements->getWithImages()) {
+            $issueDetailsLabel .= ' ' . $this->translator->trans('issues.with_images', [], 'report');
+        }
+        $elements[] = $issueDetailsLabel;
+
+        return $elements;
     }
 }
